Extract VariantHeader component
