+++ src/pages/FocusPage.tsx
@@ -1,15 +1,20 @@
 import React from 'react';
-import { VariantDashboard, Props as VariantDashboardProps } from '../components/VariantDashboard';
-import {
-  InternationalComparison,
-  Props as InternationalComparisonProps,
-} from '../components/InternationalComparison';
+import { VariantDashboard } from '../components/VariantDashboard';
+import { InternationalComparison } from '../components/InternationalComparison';
+import { Country, Variant } from '../services/api-types';
+import { VariantHeader } from '../components/VariantHeader';
 
-export type Props = VariantDashboardProps & InternationalComparisonProps;
+interface Props {
+  country: Country;
+  matchPercentage: number;
+  variant: Variant;
+}
 
 export const FocusPage = (props: Props) => {
   return (
     <>
+      <VariantHeader {...props} />
+      <hr />
       <VariantDashboard {...props} />
       <hr />
       <InternationalComparison {...props} />
